+++ Changelog.md
@@ -5,6 +5,7 @@
 - Fixed warnings in `interop` crate
 - Implemented `From` and `Into`for the conversion of a `DspVec` into a `std::Vec` to address https://github.com/liebharc/basic_dsp/issues/46
 - Increased minimum Rust version to 1.43 as the deeper integration with standard types requires https://rust-lang.github.io/rfcs/2451-re-rebalancing-coherence.html
+- Fix: in `zero_pad` and `zero_pad_b` used a different definiton for `PaddingOption::Center` if a complex vector with even points is padded to a resulting vector with odd points
 
 ## Version 0.9.0